Frontcamp React 2 New webpack config SSR
