+++ README.md
@@ -1,29 +1,19 @@
 # EPAM Frontcamp 2017 (Gomel Edition)
 
-## Hometask 3 - webpack
+## REACT Hometask 3
 **Mentee - Vitali Rusau**
 
-## Webpack
+## React Task 2
 I have 2 separate webpack configs - one for dev and second for prod.
 All source files are placed in ./src folder
 
+* React Router
+* SSR
+
 commands
-* **start** - starts Webpack with preconfigured webpack-dev-sever
+* **dev** - starts Webpack with preconfigured webpack-dev-sever
 * **build** - builds project for dev
 * **build-prod** - builds project for production
 
 
-Additional:
-* Used **google-fonts-webpack-plugin** to manage Google Web Fonts and use them locally instead of external resource
-* Used **babel-loader** plugin with preset ENV to use ES6 code features
-
-## Additional
-Installed **ESLint** (with Aitbnb config) and **jscs**
-[eslint with airbnb config](https://www.themarketingtechnologist.co/eslint-with-airbnb-javascript-style-guide-in-webstorm/)
-[jscs config](https://www.themarketingtechnologist.co/how-to-get-airbnbs-javascript-code-style-working-in-webstorm/)
-
 ## Useful links
-* [Fetch browser polyfill](https://www.npmjs.com/package/whatwg-fetch)
-* [Babel plugin: Class properties transform](http://babeljs.io/docs/plugins/transform-class-properties/)
-* [Pug – template engine for Node.js](https://github.com/pugjs/pug)
-* [Winston logger](https://github.com/winstonjs/winston)
