+++ src/client/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '../js/appComponent';
+import App from '../common/appComponent';
 
 ReactDOM.render(<App/>, document.getElementById('root'));
