+++ src/common/appComponent.js
