+++ src/common/components/AuthorItem/AuthorItem.js
