--- webpack.config.js
@@ -1,1 +1,0 @@
-module.exports = (process.env.NODE_ENV === 'production') ? require('./webpack.config.prod') : require('./webpack.config.dev');
