+++ webpack/options.js
@@ -1,8 +1,8 @@
 const path = require('path');
 
 const OPTIONS = {
-    distPath: path.resolve(__dirname, 'dist'),
-    publicPath: path.resolve(__dirname, 'public'),
+    publicPath: path.resolve(__dirname, '../public'),
+    serverBuildPath: path.resolve(__dirname, '../build'),
     cssOutputFilename: 'style.css',
     cssOutputFilenameProd: 'style.min.css'
 };
